+++ packages/npm/less-plugins/packages/css-logical/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.7
+
+- update export name
+
 ## 0.0.6
 
 - update replace reg