+++ packages/npm/less-plugins/packages/rtl/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.5
+
+- update rtl to fix left and right
+
+- Update dependencies @less-plugins/css-logical@0.0.7
+
 ## 0.0.4
 
 - Update dependencies @less-plugins/replace-properties@0.0.4